Data Model API

The genser data model is the intermediate representation between the application data and types serializable by a serialization format.

Base Types

Genser.GenserDataType — Type

Genser.GenserDataType{tag}

Base Genser data model type.

Genser.GenserValue — Type

Genser.GenserValue{V, tag}

Base Genser data model type for simple values.

Category Types

Genser.GenserNothingValue — Type

Genser.GenserNothingValue{tag}

A category type for genser values that hold a Nothing.

Genser.GenserNumberValue — Type

Genser.GenserNumberValue{V <: Number, tag}

A category type for genser values that hold a Number.

Genser.GenserRealValue — Type

Genser.GenserRealValue{V <: Real, tag}

A category type for genser values that hold a Real.

Genser.GenserIntegerValue — Type

Genser.GenserIntegerValue{V <: Integer, tag}

A category type for genser values that hold an Integer.

Genser.GenserSignedValue — Type

Genser.GenserSignedValue{V <: Signed, tag}

A category type for genser values that hold a Signed.

Genser.GenserUnsignedValue — Type

Genser.GenserUnsignedValue{V <: Unsigned, tag}

A category type for genser values that hold an Unsigned.

Genser.GenserFloatValue — Type

Genser.GenserFloatValue{V <: AbstractFloat, tag}

A category type for genser values that hold an AbstractFloat.

Genser.GenserStringValue — Type

Genser.GenserStringValue{tag}

A category type for genser values that hold an AbstractString.

Genser.GenserBinaryValue — Type

Genser.GenserBinaryValue{E <: Encoding}(value::Vector{UInt8})

A category type for genser values that hold binary data. The encoding type parameter is used to convert the data to and from a string representation for formats that do not support binary types.

Nothing Types

Genser.GenserUndefined — Type

Genser.GenserUndefined()

A type that holds nothing, tagged to suggest that it holds an "undefined" value.

Genser.GenserNull — Type

Genser.GenserNull()

A type that holds nothing, tagged to suggest that it holds a "null" value.

Value Types

Genser.GenserAny — Type

Genser.GenserAny(value::GenserDataType)

A type that may or may hold any type of value.

The is the default type for the Any type.

Genser.GenserInt8 — Type

Genser.GenserInt8(value::Int8)

A type that holds an Int8 value.

Genser.GenserUInt8 — Type

Genser.GenserUInt8(value::UInt8)

A type that holds a UInt8 value.

Genser.GenserInt16 — Type

Genser.GenserInt16(value::Int16)

A type that holds an Int16 value.

Genser.GenserUInt16 — Type

Genser.GenserUInt16(value::UInt16)

A type that holds a UInt16 value.

Genser.GenserInt32 — Type

Genser.GenserInt32(value::Int32)

A type that holds an Int32 value.

Genser.GenserUInt32 — Type

Genser.GenserUInt32(value::Int32)

A type that holds a UInt32 value.

Genser.GenserInt64 — Type

Genser.GenserInt64(value::Int64)

A type that holds an Int64 value.

Genser.GenserUInt64 — Type

Genser.GenserUInt64(value::UInt64)

A type that holds a UInt64 value.

Genser.GenserInt128 — Type

Genser.GenserInt128(value::Int128)

A type that holds an Int128 value.

Genser.GenserUInt128 — Type

Genser.GenserUInt128(value::UInt128)

A type that holds a UInt128 value.

Genser.GenserBigInt — Type

Genser.GenserBigInt(value::BigInt)

A type that holds a BigInt value.

Genser.GenserFloat16 — Type

Genser.GenserFloat16(value::Float16)

A type that holds a Float16 value.

Genser.GenserFloat32 — Type

Genser.GenserFloat32(value::Float32)

A type that holds a Float32 value.

Genser.GenserFloat64 — Type

Genser.GenserFloat64(value::Float64)

A type that holds a Float64 value.

Genser.GenserBigFloat — Type

Genser.GenserBigFloat(value::BigFloat)

A type that holds a BigFloat value.

Genser.GenserRational — Type

Genser.GenserRational(value::Rational)

A type that holds a Rational value.

Genser.GenserBool — Type

Genser.GenserBool(value::Bool)

A type that holds a Bool value.

Genser.GenserChar — Type

Genser.GenserChar(value::Char)

A type that holds a Char value.

Genser.GenserSymbol — Type

Genser.GenserSymbol(value::Symbol)

A type that holds a Symbol value.

Typically, a Symbol is serialized into its string representation, but formats may have built-in support for type type.

Genser.GenserUUID — Type

Genser.GenserUUID(value::Base.UUID)

A type that holds a UUID value.

Typically, a UUID is serialized into its string representation for text-based formats and binary representation for binary formats. Some formats may have built-in support for the type.

Genser.GenserString — Type

Genser.GenserString(value::AbstractString)

A type that holds a String value.

Genser.GenserURI — Type

Genser.GenserURI(value::AbstractString)

A type that holds a String value tagged as a URI value.

Genser.GenserBinary — Type

Genser.GenserBinary(value::Vector{UInt8})

A binary type with no string encoding.

This is the default Genser type for Vector{UInt8} value.

Complex Types

Genser.GenserSequence — Type

Genser.GenserSequence{V <: GenserDataType}(AbstractVector{V})

A type that holds an AbstractVector of values.

Genser.GenserSet — Type

Genser.GenserSet{V <: GenserDataType}(AbstractSet{V})

A type that holds an AbstractSet of values.

Typically, a set is serialized in the same format as a sequence.

Genser.GenserTuple — Type

Genser.GenserSet{V <: Tuple}(value::V)

A type that holds a Tuple of values.

Typically, a tuple is serialized in the same format as a sequence.

Genser.GenserDict — Type

Genser.GenserDict{K <: GenserDataType, V <: GenserDataType}(value::AbstractDict{K,V}())

A type that holds an AbstractDict of values.

For compatibility, the K type should be type that can be stringified.

Genser.GenserRecord — Type

Genser.GenserRecord{V <: NamedTuple}(value::V)

A type that holds a fixed set of key-value pairs.

The is the default type for all structs.

Genser.GenserOptional — Type

Genser.GenserOptional{V <: GenserDataType}(value::Union{GenserUndefined, V})

A type that may or may not hold a value.

The is the default type for all Union types that include Nothing.

Genser.GenserVariant — Type

Genser.GenserVariant{V <: GenserDataType}(value::V)

A type that may contain a value of one or more types.

The V type parameter is always a Union type.

The is the default type for all Union types that do not include Nothing.

Metadata Types

Genser.Tag — Type

Genser.Tag{tag}

The type of genser tags.

Tags provide additional semantic metadata about a genser type. They are particular useful for string values because some (de)serializers are capable of including additional metadata in their output.

Each tag is a singleton of a unique data type.

Genser.Tag — Method

Genser.Tag(tag)

Construct a tag for a Genser type.

Genser.Encoding — Type

Genser.Encoding{encoding}

Encoding for binary values when converted to and from string values.

Genser.Encoding — Method

Genser.Encoding(encoding)

Construct an encoding.

Functions

Genser.gensertypefor — Function

Genser.gensertypefor(T::Type) :: Type{<: GenserDataType}

Get the Genser data model type for a Julia type.

Genser.tag — Function

Genser.tag(v) :: Genser.Tag

Get the tag of a Genser data type or Genser value.

Genser.encoding — Function

Genser.encoding(::Type{<: GenserDataType}) :: Genser.Encoding

Get the string encoding of a genser data type or value.

Macros

Genser.@fieldtype — Macro

Genser.@fieldtype(type, fieldkey, gensertype)

Override the genser type for a struct field.

using Genser

struct Person
    id::Base.UUID
    name::String
end

# Always convert the UUID to a string
@fieldtype Persion :id GenserString

Genser.@fieldencoding — Macro

Genser.@fieldencoding(type, fieldkey, encodingkey)

Set the string encoding for a binary value.

using Genser

struct DataContainer
    contents::Vector{UInt8}
end

# Encode as base64 when serialized as a string
@fieldencoding DataContainer :contents :base64